requirements.txt
pydicom>=2.4.0
watchdog>=3.0.0
rich>=13.0.0
requests>=2.28.0
python-dotenv>=1.0.0
streamlit>=1.28.0
